refactor: migrate Utterances to ts
