--- src/components/Utterances.bs.js
@@ -1,81 +1,0 @@
-// Generated by ReScript, PLEASE EDIT WITH CARE
-
-import * as List from "bs-platform/lib/es6/list.js";
-import * as React from "react";
-import * as Caml_option from "bs-platform/lib/es6/caml_option.js";
-
-function append(child, parent) {
-  if (parent !== undefined) {
-    Caml_option.valFromOption(parent).appendChild(child);
-    return ;
-  }
-  
-}
-
-function Utterances(Props) {
-  var utterancesRef = React.useRef(null);
-  React.useEffect(function () {
-        var utterances = document.createElement("script");
-        List.iter((function (param) {
-                utterances.setAttribute(param[0], param[1]);
-                
-              }), {
-              hd: [
-                "src",
-                "https://utteranc.es/client.js"
-              ],
-              tl: {
-                hd: [
-                  "repo",
-                  "sehyunchung/sehyunchung.dev"
-                ],
-                tl: {
-                  hd: [
-                    "issue-term",
-                    "pathname"
-                  ],
-                  tl: {
-                    hd: [
-                      "theme",
-                      "photon-dark"
-                    ],
-                    tl: {
-                      hd: [
-                        "crossorigin",
-                        "anonymous"
-                      ],
-                      tl: {
-                        hd: [
-                          "async",
-                          "true"
-                        ],
-                        tl: /* [] */0
-                      }
-                    }
-                  }
-                }
-              }
-            });
-        append(utterances, Caml_option.nullable_to_opt(utterancesRef.current));
-        return (function (param) {
-                  
-                });
-      });
-  return React.createElement("div", {
-              ref: utterancesRef,
-              className: "utterances",
-              style: {
-                margin: "20px 0",
-                maxWidth: "100%"
-              }
-            });
-}
-
-var make = Utterances;
-
-export {
-  append ,
-  make ,
-  
-}
-/* react Not a pure module */
